Align @external with @shareable for how it behaves on a type.
i.e. marking the type external makes all fields external.

Fixes #2156
